code2/requirements.txt
# FastAPI for web server
fastapi
uvicorn

# OpenAI for LLM
openai

# Environment variables
python-dotenv
